pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gitask"
version = "1.1.0"
description = "A CLI tool to streamline your workflow by integrating with your PMT and VCS."
readme = "README.md"
requires-python = ">=3.6"
license = {text = "MIT License"}
authors = [
    {name = "Shahaf Mordechay"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "click",
    "jira",
    "requests",
    "python-gitlab>=5.6.0",
    "PyGithub>=2.1.1",
    "pyjwt[crypto]>=2.4.0"
]

[project.urls]
homepage = "https://github.com/shahafMordechay/Gitask"
repository = "https://github.com/shahafMordechay/Gitask"


[project.scripts]
gitask = "gitask.main:cli"

[tool.pytest.ini_options]
addopts = "-ra -q"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
